REFACTOR : contentObserver 모듈화
