+++ domain/src/main/java/com/fakedevelopers/domain/repository/ImageRepository.kt
@@ -4,10 +4,12 @@ import com.fakedevelopers.domain.model.AlbumItem
 import com.fakedevelopers.domain.model.MediaInfo
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
 
 interface ImageRepository {
     fun isValid(uri: String): Boolean
     suspend fun getImages(path: String?): List<AlbumItem>
+    fun getImageObserver(): Flow<String>
     fun getValidUris(uris: List<String>): List<String>
     fun getDateModifiedByUri(uri: String): AlbumItem?
     suspend fun getBytesByUri(uri: String, dispatcher: CoroutineDispatcher = Dispatchers.IO): ByteArray?
